+++ src/connectProvider/web3Provider/web3-provider/events.js
@@ -3,5 +3,6 @@ export default {
   SHOW_TX_CONFIRM_MODAL: 'showTxConfirmModal',
   SHOW_MSG_CONFIRM_MODAL: 'showMessageConfirmModal',
   SHOW_TX_SIGN_MODAL: 'showTxSignModal',
-  WALLET_NOT_CONNECTED: 'walletNotConnected'
+  WALLET_NOT_CONNECTED: 'walletNotConnected',
+  ERROR_NOTIFY: 'errorNotify'
 };